Remove the CGLayer code path from MMCoreTextView.
Relying on drawRect: being able to draw on top of the existing content
works if the window's content view is opaque. In a normal, titled window
it is, but in a borderless window it is not by default.

This change gives MMFullScreenWindow an opaque contentView and removes
the CGLayer code path. It shouldn't result in any functional changes,
but on my machine it saves ~20MB of RAM per full screen window.

The Apple docs say that opaque views (like MMCoreTextView) "must totally
fill the specified rectangle with opaque content", so it's still not
correct to just draw changed content. Layer-backed views do not work
this way at all. However, since the non-full screen case relies on this
quirk of AppKit, it might make sense to only use that path until it can
be replaced with something better.