Stateful, fast renderer for MMCoreTextView
This change makes `MMCoreTextView` apply draw commands to an array of
structs that represent the state of each character cell. It marks
changed cells as needing display, and drawRect: draws from the array.

In my own tests, this renderer seems to be roughly twice as fast as the
current one. Because it seems to work for the use cases of all of the
current renderers and duplicates a lot of code, this PR replaces the
other paths instead of adding a preference.

Fixes #796.